Move the main function to Main.cpp from Sazabi.cpp Add dependency on .obj files Add a sample test
