+++ Sazabi.sln
@@ -33,14 +33,14 @@ Global
 		{83410426-0B15-352C-4BC5-CCD2A7EF1930}.Release|Win32.Build.0 = R64_CSG|Win32
 		{83410426-0B15-352C-4BC5-CCD2A7EF1930}.Release|x64.ActiveCfg = R64_CSG|Win32
 		{83410426-0B15-352C-4BC5-CCD2A7EF1930}.Release|x64.Build.0 = R64_CSG|Win32
-		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.D64_CSG|Win32.ActiveCfg = Debug|Win32
-		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.D64_CSG|Win32.Build.0 = Debug|Win32
+		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.D64_CSG|Win32.ActiveCfg = Release|Win32
+		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.D64_CSG|Win32.Build.0 = Release|Win32
 		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.D64_CSG|x64.ActiveCfg = Debug|x64
 		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.D64_CSG|x64.Build.0 = Debug|x64
 		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.Debug|Win32.ActiveCfg = Debug|Win32
 		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.Debug|Win32.Build.0 = Debug|Win32
-		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.Debug|x64.ActiveCfg = Debug|x64
-		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.Debug|x64.Build.0 = Debug|x64
+		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.Debug|x64.ActiveCfg = Release|Win32
+		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.Debug|x64.Build.0 = Release|Win32
 		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.R64_CSG|Win32.ActiveCfg = Release|Win32
 		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.R64_CSG|Win32.Build.0 = Release|Win32
 		{80871372-2D7E-43AF-9EBB-B6A77ED3BE21}.R64_CSG|x64.ActiveCfg = Release|x64
