+++ test/SazabiSpec/SazabiSpec.vcxproj.filters
@@ -1,0 +1,20 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <ClCompile Include="pch.cpp" />
+    <ClCompile Include="test_sbcommon.cpp">
+      <Filter>SB</Filter>
+    </ClCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="pch.h" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Filter Include="SB">
+      <UniqueIdentifier>{c5c91dca-edee-4831-ad5d-3dc17466180b}</UniqueIdentifier>
+    </Filter>
+  </ItemGroup>
+</Project>
