--- test/SazabiSpec/test.cpp
@@ -1,10 +1,0 @@
-#include "pch.h"
-#include "..\..\StdAfx.h"
-#include "..\..\Sazabi.h"
-#include "..\..\client_handler.h"
-
-TEST(TestCaseName, TestName) {
-  EXPECT_EQ(1, 1);
-  ClientHandler handler;
-  EXPECT_TRUE(true);
-}
